move arms to give bernoulli rewards

+++ src/main/kotlin/bandits/strategies/AbstractStrategy.kt
@@ -3,7 +3,7 @@ package bandits.strategies
 import bandits.BanditArm
 
 abstract class AbstractStrategy {
-    abstract fun updateStrategy(reward: Double, arm: BanditArm)
+    abstract fun updateStrategy(reward: Int, arm: BanditArm)
     abstract fun resetStrategy()
     abstract fun pickArm(): Int
 }
